Use GtkFileChooserNative over GtkFileChooserDialog
Currently, the app is using GtkFileChooserDialog, which requires
explicit access to the system folders. This permission makes GNOME
Software say that the app is not secure.

GtkFileChooserNative uses FreeDesktop portals, so it does not have full
access to all the folders, but rather only the files which it is given
permission by the user.

Portals will also handle recent files, folders opened recently by the
app, which are very handy features.

This dialog will eventually be deprecated too (in 4.10) by FileDialog,
which uses portals too, but by the moment I think it's better if the app
uses as less permissions as possible :)